requirements.txt
torch>=1.7.1
numpy
pandas
scikit-learn
scipy
pyyaml
notebook
tqdm
loguru
matplotlib
